Restructure genmd
